requirements.txt
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.9.1
dj-database-url==0.5.0
django-on-heroku==1.1.2
gunicorn==20.1.0
mixer==7.2.0